tests/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.3.1
cycler==0.11.0
Flask==3.1.0
fonttools==4.55.3
idna==3.7
iniconfig==1.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numexpr==2.10.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==25.1
pluggy==1.5.0
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
requests==2.32.3
seaborn==0.13.2
setuptools==72.1.0
six==1.17.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
